File Extension Redirecting For Frisk's Choice
Adding file extension redirects when using Frisk's choice in opening a
file. This will allow users to override a particular file type to be
opened as if it were a different file type.
Currently no user interface for this; it must be added by hand.
Example .conf field:
"extensionRedirects": [{
"ext": ".rb",
"redirectExt": ".txt"
}]